fix(parser.js): Allow all variants of directive.
Due to many changes to the RFC, there are currently four versions of the
'Acknowledgements' directive. Pluralisation + American/British. Current
RFC draft states 'Acknowledgments'. This fixes a previous fix which
attempts to handle lack of pluralisation. It only corrected American
English because it was modifying a regular expression to normalise
American English.